backend/services/scraper_service.py
import requests
from bs4 import BeautifulSoup
from typing import Dict, Optional
import re
from datetime import datetime


class ScraperService:
    """Service for scraping team statistics from NCAA and conference websites"""
    
    def __init__(self):
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }
    
    def scrape_team_stats(
        self,
        team_name: str,
        sport: str,
        date_range: Optional[Dict] = None
    ) -> Dict:
        """
        Scrape team statistics from various sources
        
        Args:
            team_name: Name of the team to scrape
            sport: Type of sport (soccer or baseball)
            date_range: Optional date range for stats
            
        Returns:
            Dictionary containing scraped statistics
        """
        
        stats = {
            "team_name": team_name,
            "sport": sport,
            "date_range": date_range,
            "sources": []
        }
        
        try:
            # Try NCAA.org
            ncaa_stats = self._scrape_ncaa(team_name, sport, date_range)
            if ncaa_stats:
                stats["ncaa_stats"] = ncaa_stats
                stats["sources"].append("NCAA.org")
        except Exception as e:
            print(f"Error scraping NCAA: {str(e)}")
        
        try:
            # Try Big West Conference
            bigwest_stats = self._scrape_bigwest(team_name, sport, date_range)
            if bigwest_stats:
                stats["conference_stats"] = bigwest_stats
                stats["sources"].append("Big West Conference")
        except Exception as e:
            print(f"Error scraping Big West: {str(e)}")
        
        try:
            # Try to get UCSB comparison data if available
            if "ucsb" in team_name.lower() or "santa barbara" in team_name.lower():
                ucsb_stats = self._scrape_ucsb_athletics(sport, date_range)
                if ucsb_stats:
                    stats["ucsb_comparison"] = ucsb_stats
                    stats["sources"].append("UCSB Athletics")
        except Exception as e:
            print(f"Error scraping UCSB: {str(e)}")
        
        # Add mock data for demonstration purposes
        if sport == "soccer":
            stats["summary"] = self._get_mock_soccer_stats(team_name)
        else:  # baseball
            stats["summary"] = self._get_mock_baseball_stats(team_name)
        
        return stats
    
    def _scrape_ncaa(
        self,
        team_name: str,
        sport: str,
        date_range: Optional[Dict]
    ) -> Optional[Dict]:
        """Scrape statistics from NCAA.org"""
        
        # NCAA.org scraping logic
        # Note: Actual implementation would need to handle NCAA's specific HTML structure
        # and potentially use their API if available
        
        base_url = "https://www.ncaa.org"
        # This is a placeholder - actual URL structure depends on NCAA's website
        
        try:
            # Example structure - would need to be adapted to actual NCAA website
            response = requests.get(base_url, headers=self.headers, timeout=10)
            if response.status_code == 200:
                soup = BeautifulSoup(response.content, 'html.parser')
                # Parse relevant statistics
                return {
                    "source": "NCAA.org",
                    "scraped_at": datetime.now().isoformat(),
                    "data": "NCAA data would be parsed here"
                }
        except Exception as e:
            print(f"NCAA scraping error: {str(e)}")
            return None
    
    def _scrape_bigwest(
        self,
        team_name: str,
        sport: str,
        date_range: Optional[Dict]
    ) -> Optional[Dict]:
        """Scrape statistics from Big West Conference website"""
        
        base_url = "https://bigwest.org"
        
        try:
            response = requests.get(base_url, headers=self.headers, timeout=10)
            if response.status_code == 200:
                soup = BeautifulSoup(response.content, 'html.parser')
                return {
                    "source": "Big West Conference",
                    "scraped_at": datetime.now().isoformat(),
                    "data": "Conference data would be parsed here"
                }
        except Exception as e:
            print(f"Big West scraping error: {str(e)}")
            return None
    
    def _scrape_ucsb_athletics(
        self,
        sport: str,
        date_range: Optional[Dict]
    ) -> Optional[Dict]:
        """Scrape statistics from UCSB Athletics website"""
        
        base_url = "https://ucsbgauchos.com"
        
        try:
            response = requests.get(base_url, headers=self.headers, timeout=10)
            if response.status_code == 200:
                soup = BeautifulSoup(response.content, 'html.parser')
                return {
                    "source": "UCSB Athletics",
                    "scraped_at": datetime.now().isoformat(),
                    "data": "UCSB data would be parsed here"
                }
        except Exception as e:
            print(f"UCSB Athletics scraping error: {str(e)}")
            return None
    
    def _get_mock_soccer_stats(self, team_name: str) -> Dict:
        """Return mock soccer statistics for demonstration"""
        current_year = datetime.now().year
        return {
            "team": team_name,
            "season": str(current_year),
            "record": {
                "wins": 12,
                "losses": 5,
                "draws": 3,
                "win_percentage": 0.675
            },
            "offense": {
                "goals_per_game": 2.1,
                "shots_per_game": 14.5,
                "shot_accuracy": 0.51,
                "corner_kicks_per_game": 6.2,
                "possession_percentage": 54.3
            },
            "defense": {
                "goals_against_per_game": 1.3,
                "shots_against_per_game": 11.2,
                "saves_per_game": 4.8,
                "clean_sheets": 7,
                "tackle_success_rate": 0.68
            },
            "discipline": {
                "yellow_cards": 32,
                "red_cards": 2,
                "fouls_per_game": 12.3
            },
            "key_players": [
                {
                    "name": "Forward #10",
                    "position": "Forward",
                    "goals": 8,
                    "assists": 5,
                    "shots_per_game": 3.2
                },
                {
                    "name": "Midfielder #7",
                    "position": "Midfielder", 
                    "goals": 4,
                    "assists": 7,
                    "pass_accuracy": 0.84
                }
            ]
        }
    
    def _get_mock_baseball_stats(self, team_name: str) -> Dict:
        """Return mock baseball statistics for demonstration"""
        return {
            "team": team_name,
            "record": {
                "wins": 28,
                "losses": 15,
                "win_percentage": 0.651
            },
            "hitting": {
                "batting_average": 0.287,
                "on_base_percentage": 0.362,
                "slugging_percentage": 0.445,
                "home_runs": 42,
                "runs_per_game": 6.2,
                "strikeouts_per_game": 7.8,
                "stolen_bases": 34,
                "stolen_base_percentage": 0.773
            },
            "pitching": {
                "team_era": 3.84,
                "whip": 1.32,
                "strikeouts_per_9": 9.2,
                "walks_per_9": 3.4,
                "home_runs_per_9": 1.1,
                "save_percentage": 0.82
            },
            "fielding": {
                "fielding_percentage": 0.971,
                "errors": 38,
                "double_plays": 32
            },
            "key_players": [
                {
                    "name": "Pitcher #21",
                    "position": "Starting Pitcher",
                    "era": 2.94,
                    "wins": 7,
                    "strikeouts": 82,
                    "whip": 1.15
                },
                {
                    "name": "Outfielder #5",
                    "position": "Outfielder",
                    "batting_avg": 0.341,
                    "home_runs": 12,
                    "rbi": 38,
                    "obp": 0.412
                }
            ]
        }
