backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
openai>=1.60.0
httpx>=0.27.0,<0.28.0
gunicorn==21.2.0
beautifulsoup4==4.12.3
lxml==5.1.0
